NTBS-960: Move the rest of notification display methods to partial classes
