+++ ntbs-service/Models/Entities/ImmunosuppressionDetails.cs
@@ -1,6 +1,4 @@
-﻿
-using System.ComponentModel.DataAnnotations;
-using System.Text;
+﻿using System.ComponentModel.DataAnnotations;
 using EFAuditer;
 using ExpressiveAnnotations.Attributes;
 using Microsoft.EntityFrameworkCore;
@@ -10,9 +8,9 @@ using ntbs_service.Models.Validations;
 namespace ntbs_service.Models.Entities
 {
     [Owned]
-    public class ImmunosuppressionDetails : ModelBase, IOwnedEntityForAuditing
+    public partial class ImmunosuppressionDetails : ModelBase, IOwnedEntityForAuditing
     {
-        [AssertThat(@"TestAtLeastOneSelectedWhenYes", 
+        [AssertThat(nameof(TestAtLeastOneSelectedWhenYes), 
             ErrorMessage = ValidationMessages.ImmunosuppressionTypeRequired)]
         public Status? Status { get; set; }
 
@@ -26,29 +24,7 @@ namespace ntbs_service.Models.Entities
         [RegularExpression(ValidationRegexes.CharacterValidation, ErrorMessage = ValidationMessages.StandardStringFormat)]
         [Display(Name = "Immunosuppression type description")]
         public string OtherDescription { get; set; }
-
-        public string CreateTypesOfImmunosuppressionString()
-        {
-            var sb = new StringBuilder();
-            sb.Append(HasBioTherapy == true ? "Biological Therapy" : string.Empty);
-
-            if (HasTransplantation == true)
-            {
-                if (sb.Length != 0)
-                    sb.Append(", ");
-                sb.Append("Transplantation");
-            }
-
-            if (HasOther == true)
-            {
-                if (sb.Length != 0)
-                    sb.Append(", ");
-                sb.Append("Other");
-            }
-
-            return sb.ToString();
-        }
-
+        
         string IOwnedEntityForAuditing.RootEntityType => RootEntities.Notification;
 
         public bool TestAtLeastOneSelectedWhenYes
