+++ ntbs-service/Models/Entities/MDRDetails.cs
@@ -9,7 +9,7 @@ using ntbs_service.Models.Validations;
 namespace ntbs_service.Models.Entities
 {
     [Owned]
-    public class MDRDetails : ModelBase, IOwnedEntityForAuditing
+    public partial class MDRDetails : ModelBase, IOwnedEntityForAuditing
     {
         [Display(Name = "Has the patient been exposed to a known RR/MDR/XDR-TB Case?")]
         public Status? ExposureToKnownCaseStatus { get; set; }
