+++ ntbs-service/Models/Entities/MDRDetailsDisplay.cs
@@ -1,0 +1,7 @@
+﻿namespace ntbs_service.Models.Entities
+{
+    public partial class MDRDetails
+    {
+        public string MDRCaseCountryName => Country?.Name;
+    }
+}
