+++ ntbs-service/Models/Entities/PatientTBHistoryDisplay.cs
@@ -1,0 +1,10 @@
+﻿using ntbs_service.Helpers;
+
+namespace ntbs_service.Models.Entities
+{
+    public partial class PatientTBHistory
+    {
+        public string PreviouslyHadTBYesNo => (PreviouslyHadTB).FormatYesNo();
+
+    }
+}
