+++ ntbs-service/Models/Entities/TravelDetails.cs
@@ -9,7 +9,7 @@ using ntbs_service.Models.Validations;
 namespace ntbs_service.Models.Entities
 {
     [Owned]
-    public class TravelDetails : ModelBase, ITravelOrVisitorDetails, IOwnedEntityForAuditing
+    public partial class TravelDetails : ModelBase, ITravelOrVisitorDetails, IOwnedEntityForAuditing
     {
         private const int MaxTotalLengthOfStay = 24;
