+++ ntbs-service/Models/Entities/TravelDetailsDisplay.cs
@@ -1,0 +1,9 @@
+﻿using ntbs_service.Helpers;
+
+namespace ntbs_service.Models.Entities
+{
+    public partial class TravelDetails
+    {
+        public string HasRecentTravel => HasTravel.FormatYesNo();
+    }
+}
