+++ ntbs-service/Models/Entities/VisitorDetailsDisplay.cs
@@ -1,0 +1,9 @@
+﻿using ntbs_service.Helpers;
+
+namespace ntbs_service.Models.Entities
+{
+    public partial class VisitorDetails
+    {
+        public string HasRecentVisitor => HasVisitor.FormatYesNo();
+    }
+}
