+++ ntbs-service/Pages/Notifications/OverviewPartials/_ComorbiditiesAndImmunosuppresionOverviewPartial.cshtml
@@ -48,7 +48,7 @@
         </nhs-grid-column>
         <nhs-grid-column grid-column-width="OneQuarter">
             <div class="notification-details-label"> Type </div>
-            <div class="cell-min-height">@Model.Notification.ImmunosuppressionDetails.CreateTypesOfImmunosuppressionString()</div>
+            <div class="cell-min-height">@Model.Notification.ImmunosuppressionDetails.FormattedTypesOfImmunosuppression()</div>
         </nhs-grid-column>
         <nhs-grid-column grid-column-width="OneHalf">
             <div class="notification-details-label"> Other immunosuppression </div>
