+++ ntbs-service/Pages/Notifications/OverviewPartials/_HospitalDetailsOverviewPartial.cshtml
@@ -43,11 +43,11 @@
         <nhs-grid-row>
             <nhs-grid-column grid-column-width="OneQuarter">
                 <div class="notification-details-label"> Denotification Date </div>
-                <div class="cell-min-height"> @Model.Notification.FormattedDenotificationDate </div>
+                <div class="cell-min-height"> @Model.Notification?.DenotificationDetails.FormattedDenotificationDate </div>
             </nhs-grid-column>
             <nhs-grid-column grid-column-width="OneQuarter">
                 <div class="notification-details-label"> Denotification Reason </div>
-                <div class="cell-min-height" id="overview-denotification-reason"> @Model.Notification.DenotificationReasonString </div>
+                <div class="cell-min-height" id="overview-denotification-reason"> @Model.Notification?.DenotificationDetails.DenotificationReasonString </div>
             </nhs-grid-column>
         </nhs-grid-row>
     }
