+++ ntbs-service/Pages/Notifications/OverviewPartials/_MDRDetailsOverviewPartial.cshtml
@@ -43,7 +43,7 @@
             case Status.No:
                 <nhs-grid-column grid-column-width="OneQuarter">
                     <div class="notification-details-label"> Country of contact </div>
-                    <div class="cell-min-height"> @Model.Notification.MDRCaseCountryName </div>
+                    <div class="cell-min-height"> @Model.Notification.MDRDetails.MDRCaseCountryName </div>
                 </nhs-grid-column>
                 break;
         }
