+++ ntbs-service/Pages/Notifications/OverviewPartials/_PreviousHistoryOverviewPartial.cshtml
@@ -20,7 +20,7 @@
     <nhs-grid-row>
         <nhs-grid-column grid-column-width="OneQuarter">
             <div class="notification-details-label"> Previous notification of TB </div>
-            <div class="cell-min-height"> @Model.Notification.PreviouslyHadTBYesNo </div>
+            <div class="cell-min-height"> @Model.Notification.PatientTBHistory.PreviouslyHadTBYesNo </div>
         </nhs-grid-column>
         <nhs-grid-column grid-column-width="OneQuarter">
             <div class="notification-details-label"> Year of previous diagnosis </div>
