+++ ntbs-service/Pages/Notifications/OverviewPartials/_SocialRiskFactorsOverviewPartial.cshtml
@@ -24,7 +24,7 @@
         </nhs-grid-column>
         <nhs-grid-column grid-column-width="OneHalf">
             <div class="notification-details-label"> Time periods </div>
-            <div> @Model.Notification.DrugRiskFactorTimePeriods </div>
+            <div> @Model.Notification.SocialRiskFactors.DrugRiskFactorTimePeriods </div>
         </nhs-grid-column>
     </nhs-grid-row>
     <nhs-grid-row>
@@ -34,7 +34,7 @@
         </nhs-grid-column>
         <nhs-grid-column grid-column-width="OneHalf">
             <div class="notification-details-label"> Time periods </div>
-            <div> @Model.Notification.HomelessRiskFactorTimePeriods </div>
+            <div> @Model.Notification.SocialRiskFactors.HomelessRiskFactorTimePeriods </div>
         </nhs-grid-column>
     </nhs-grid-row>
     <nhs-grid-row>
@@ -44,7 +44,7 @@
         </nhs-grid-column>
         <nhs-grid-column grid-column-width="OneHalf">
             <div class="notification-details-label"> Time periods </div>
-            <div> @Model.Notification.ImprisonmentRiskFactorTimePeriods </div>
+            <div> @Model.Notification.SocialRiskFactors.ImprisonmentRiskFactorTimePeriods </div>
         </nhs-grid-column>
     </nhs-grid-row>
