+++ ntbs-service/Pages/Shared/_AlertsAndActions.cshtml
@@ -48,7 +48,7 @@
     <nhs-grid-column grid-column-width="OneThird">
         <div class="@manageNotificationClass">
             <nhs-details display-text="Manage notification" nhs-details-type="Expander" id="manage-notification">
-                @if (Model.Notification.OverOneYearOld)
+                @if (Model.Notification.IsOverOneYearOld)
                 {
                     <form method="post">
                         <input type="hidden" value="@Model.NotificationId" name="NotificationId"/>
